[fix] Changed Ember root to div with given id to prevent conflict with Turbolinks (1h)
By default, Ember attaches the app to the `<body>` element. This would be fine except:
1. All Ember's internal event handling is attached to the rootElement
2. Turbolinks replaces the `<body>` element, even if it preserves the component `<div>` to which Ember Islands attaches

When Turbolinks replaces the `<body>`, all event handling for Ember is broken; the solution is for Ember to attach its application instance to the component `<div>` instead, which Turbolinks is able to preserve (along with its event listeners) between page loads.