[skip] Restored the "Narrow Your Search" section heading to modifiers (5m)
This was a side-effect brought to light by the changes made to allow "unlisting" certain modifiers. When preparing default content, instead of modifying _copies_ of each modifier, changing the section name, we were modifying the original objects, so the special `+` token/modifier ended up getting the changes in its modifiers, as well. But with unlisting, the general content list and the `+` modifier contained separate objects, meaning the `+`'s modifier list now _didn't_ get this behind-the-scenes modification. This PR essentially drops the indirection and assigns "Narrow your Search" as the section when generating modifiers in the first place.